Improve/fix redraw after command
In my previous commit I naively called the function to update the
tabline after the command runs. The problem there is that it totally
obliterates output from the previous command which makes sense because
right after the first command finishes, a second separate command gets
run to update the tabline. This commit fixes that through much hackery.