Update tabline using tab local variables
Prior to this commit, updating the tabline involved tabnext'ing to each
tab and calling getcwd(). With this approach it is impossible (at least
I think so) to update the tabline after exiting the command line window
(which I want to be able to do) and also I thought this method of
slogging through tabs one by one seemed inelegant.