Alter display, update tabline after :tabmove
The home directory will now display as just '~', the root directory
properly displays as '/', and for a tab in directory A, a string of
adjacent tabs AS will appear as '"' (the idea behind this was just to
conserve tabline space, we'll see how I like it). I also update the
tabline after a :tabmove command.