Make code less hacky
I was doing terrible things involving rewriting the command line history
and mapping `<CR>` in command line mode and the command line window all
so the tabline would update when commands like :lcd and :tabmove were
executed. It was a mess. Just a little while ago I learned that tpope's
flagship plugin can also put cwds on the tabline but his code was very
clean. Upon closer inspection I realized that all I had to do to get the
tabline to update like I wanted it to was to set the tabline like this:

        set tabline=%!expr

Vim will then re-evaluate expr whenever the tabline needs to get redrawn
which is fairly often apparently. I had previously been doing:

        let &tabline = expr

This meant that to automatically get the tabline to be redrawn I needed
to add autocommands like TabEnter.